+++ src/server/routes/apiv3/users.js
@@ -37,7 +37,7 @@ const validator = {};
  *          lang:
  *            type: string
  *            description: language
- *            example: 'en-US'
+ *            example: 'en_US'
  *          status:
  *            type: integer
  *            description: status